Refactor: Pages 작성 규칙 변경 및 폴더 규칙 변경 적용, Waiting 폴더 내 Component 이름 변경
